requirements.txt
Django==4.2.1
psycopg2==2.9.6
SQLAlchemy==2.0.15
djangorestframework==3.14.0
cloudinary==1.33.0
bcrypt==4.0.1
pyJWT==2.3.0